// src/interfaces/game.ts
export interface IGame {
  id: string;
  game_name: string;
  description: string;
  price: number;
  banner_image: string;
  slide_images: string[];
  logo: string;
  path_name: string;
  folder_installation: string;
  rating: number;
  count_rating: number;
  count_view: number;
  count_play: number;
  exec_file_name: string;
  created_by: string;
}

export interface IGameResponse {
  status: string;
  data: null | IGame;
  error: null | {
    statusCode: number;
    message: string;
  };
}